Fixed control bus

+++ libemulation/Core/OEComponent.cpp
@@ -63,7 +63,7 @@ bool OEComponent::postMessage(OEComponent *sender, int message, void *data)
 
 void OEComponent::notify(OEComponent *sender, int notification, void *data)
 {
-	for (int i = 0; i < observers[notification].size(); i++)
+	for (size_t i = 0; i < observers[notification].size(); i++)
 		observers[notification][i]->notify(this, notification, data);
 }
